merge

+++ src/components/ui/DishCard.jsx
@@ -3,6 +3,8 @@ import React, { useState } from 'react';
 export default function DishCard({ dish, user, like }) {
   const [isLike, setIsLike] = useState(like);
   async function likeRules() {
+    if (!user) return alert('Чтобы добавлять в избранное нужно зарегистрироваться');
+
     if (isLike) {
       const response = await fetch(`/api/likes/${user.id}/${dish.id}`, { method: 'DELETE' });
       response.status === 200 ? setIsLike(false) : setIsLike(null);
@@ -16,7 +18,14 @@ export default function DishCard({ dish, user, like }) {
   return (
     <div className="col">
       <div className="card">
-        <img src={dish.img} className="card-img-top" style={{ width: '100%' }} />
+        <div className="imgCard">
+          <img src={dish.img} className="card-img-top" style={{ width: '100%' }} />
+          {/* {user
+            ? ( */}
+          <img className="favoriteImg" onClick={() => likeRules()} src={!isLike ? '/images/likeOff.svg' : '/images/likeOn.svg'} />
+          {/* // )
+            // : null} */}
+        </div>
         <div className="card-body">
           <h3><a href={`/dishes/${dish.id}`}>{dish.name}</a></h3>
         </div>
@@ -31,13 +40,6 @@ export default function DishCard({ dish, user, like }) {
             {' '}
             {`${dish.time} минут`}
           </div>
-          {user
-            ? (
-              <button onClick={() => likeRules()} type="button" className="btn btn-primary">
-                {!isLike ? 'Добавить в избранное' : 'Удалить из избранного'}
-              </button>
-            )
-            : null}
         </div>
       </div>
     </div>
